+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `Explorer.Series.rle_id/1` to add run-length encoding (RLE) of rows.
+
 ## [v0.12.0] - 2026-07-05
 
 ### Updated